Isolated the langchain code to this branch
