--- from_scratch/Makefile
@@ -1,12 +1,0 @@
-process:
-	poetry run python src/processing.py --use_spacy --describe
-
-tokens:
-	poetry run python src/tokenizer.py
-
-embeddings:
-	poetry run python src/embeddings.py 
-
-
-search:
-	poetry run python semantic_search.py
