--- from_scratch/pyproject.toml
@@ -1,26 +1,0 @@
-[tool.poetry]
-name = "src"
-version = "0.1.0"
-description = ""
-authors = ["Kobina Brandon <kobinabrandon@proton.me>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "<3.13, >=3.9"
-requests = "^2.32.3"
-loguru = "^0.7.2"
-tqdm = "^4.66.5"
-pymupdf = "^1.24.10"
-spacy = "^3.7.6"
-pandas = "^2.2.2"
-pyarrow = "^17.0.0"
-fastparquet = "^2024.5.0"
-sentencepiece = "^0.2.0"
-pydantic-settings = "^2.6.1"
-sentence-transformers = "^3.3.1"
-numpy = "1.26.4"
-general = {path = "../general"}
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
