--- general/Makefile
@@ -1,3 +1,0 @@
-books:
-	poetry run python general/books.py
-
