+++ src/reading.py
@@ -4,7 +4,7 @@ Provides code that reads text and prepares it for chunking.
 from langchain_core.documents import Document 
 from langchain_community.document_loaders import PyPDFLoader
 
-from general.books import Book
+from src.books import Book
 from src.cleaning import remove_non_core_pages, remove_new_line_from_documents
